% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{data_example}
\alias{data_example}
\title{Simulated effects on quantitative incidence and prognosis traits}
\format{
A data frame with 10,000 rows and 5 variables:
\describe{
  \item{xbeta}{Regression coefficient on incidence}
  \item{xse}{Standard error of \code{xbeta}}
  \item{ybeta}{Regression coefficient on prognosis}
  \item{yse}{Standard error of \code{ybeta}}
  \item{yp}{P-value of the association with prognosis}
}
}
\usage{
data_example
}
\description{
A simulated dataset for 10,000 independent variables (e.g. SNPs) consisting of regression coefficients on
incidence and prognosis, with their standard errors. Among them, 5% (500 variables) have effects on
incidence only, 5% (500 variables) on prognosis only, and 5% have correlated effects on both with a correlation
coeficient of '-0.5'. The estimates are obtained from linear regression in a simulated dataset of 20,000 individuals.
}
\examples{
# Default analysis with Hedges-Olkin adjustment for regression dilution
# Does not calculate a standard error
Sh.Model <- slopehunter(dat = data_example, xbeta_col="xbeta", xse_col="xse",
                        ybeta_col="ybeta", yse_col="yse", yp_col="yp",
                        comp.size = seq(0.03, 0.10, 0.01), xp.thresh = 0.1, coef.diff = 1,
                        correct.reg.dill = TRUE, show_adjustments = TRUE, seed=2019)

# [1] "Estimated slope: -0.258701852489861"
# [1] "SE of the slope: 0.0143848328056302"
# [1] "95\% CI: -0.286896124788896, -0.230507580190826"

Sh.Model$Sh.b
# [1] -0.2587019

#' Adj <- Sh.Model$Estimates
head(Adj$ybeta)
# [1] -0.0092889266  0.0005575032  0.0112203795 -0.0095533069  0.0082635203  0.0026550045
head(Adj$ybeta.Adj)
# [1] -0.011184217  0.004265600  0.008185743 -0.008299934  0.001626433  0.006175507
plot(Sh.Model, what = "clusters")
plot(Sh.Model, what = "classification")
}
\keyword{datasets}
