% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD_prune.R
\name{download_plink}
\alias{download_plink}
\title{Check and download PLINK 1.90 executable suitable for the operating system, and return its path
Inspired by https://github.com/MRCIEU/genetics.binaRies}
\usage{
download_plink()
}
\description{
Check and download PLINK 1.90 executable suitable for the operating system, and return its path
Inspired by https://github.com/MRCIEU/genetics.binaRies
}
