% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SH-utils.R
\name{Euclid.dist}
\alias{Euclid.dist}
\title{Calculate Euclidean distance from origin}
\usage{
Euclid.dist(x, y, ydominant = FALSE)
}
\arguments{
\item{x}{Vector of x-axis values}

\item{y}{Vector of y-axis values}

\item{ydominant}{Logical, if TRUE, larger influence will be given to the y-axis values}
}
\value{
Vector of distances
}
\description{
Calculate Euclidean distance from origin
}
