% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise_effects.R
\name{harmonise_effects}
\alias{harmonise_effects}
\title{Harmonise and format data for Slope-Hunter}
\usage{
harmonise_effects(
  incidence_dat,
  prognosis_dat,
  incidence_formatted = TRUE,
  prognosis_formatted = TRUE,
  by.pos = FALSE,
  pos_cols = c("POS.incidence", "POS.prognosis"),
  snp_cols = c("SNP", "SNP"),
  beta_cols = c("BETA.incidence", "BETA.prognosis"),
  se_cols = c("SE.incidence", "SE.prognosis"),
  EA_cols = c("EA.incidence", "EA.prognosis"),
  OA_cols = c("OA.incidence", "OA.prognosis")
)
}
\arguments{
\item{incidence_dat}{data.frame with data for incidence. It is recommended to be an output from \code{read_incidence}. If not, it tries to format it before harmonisation.}

\item{prognosis_dat}{data.frame with data for prognosis. It is recommended to be an output from \code{read_prognosis}. If not, it tries to format it before harmonisation.}

\item{incidence_formatted}{Logical indicationg whether \code{incidence_dat} is formatted using \code{read_incidence}.}

\item{prognosis_formatted}{Logical indicationg whether \code{prognosis_dat} is formatted using \code{read_prognosis}.}

\item{by.pos}{Logical, if \code{TRUE} the harmonisation will be performed by matching the exact SNP positions between the incidence and prognosis datasets.}

\item{pos_cols}{A vector of length 2 specifying the name of the genetic position columns in the incidence and prognosis datasets respectively.}

\item{snp_cols}{A vector of length 2 specifying the name of the snp columns in the incidence and prognosis datasets respectively. This is the column on which the data will be merged if \code{by.pos} is \code{FASLE}.}

\item{beta_cols}{A vector of length 2 specifying the name of the beta columns in the incidence and prognosis datasets respectively.}

\item{se_cols}{A vector of length 2 specifying the name of the se columns in the incidence and prognosis datasets respectively.}

\item{EA_cols}{A vector of length 2 specifying the name of the effect allele columns in the incidence and prognosis datasets respectively.}

\item{OA_cols}{A vector of length 2 specifying the name of the non-effect allele columns in the incidence and prognosis datasets respectively.}
}
\value{
A data.frame with harmonised effects and alleles
}
\description{
Harmonise the alleles and effects between the incidence and prognosis
}
\details{
In order to perform Slope-Hunter analysis the effect of a SNP on an incidence and prognosis traits must be harmonised to be
relative to the same allele.

This function will try to harmonise the incidence and prognosis data sets on the specified columns. Where necessary, correct strand
for non-palindromic SNPs (i.e. flip the sign of effects so that the effect allele is the same in both datasets), and drop all palindromic SNPs from the analysis (i.e. with the allele A/T or G/C).
The alleles that do not match between data sets (e.g T/C in one data set and A/C in the other) will also be dropped.
}
