% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hunt.R
\name{hunt}
\alias{hunt}
\title{Estimate collider bias}
\usage{
hunt(
  dat,
  snp_col = "SNP",
  xbeta_col = "BETA.incidence",
  xse_col = "SE.incidence",
  xp_col = "Pval.incidence",
  ybeta_col = "BETA.prognosis",
  yse_col = "SE.prognosis",
  yp_col = "Pval.prognosis",
  xp_thresh = 0.001,
  init_pi = 0.6,
  init_sigmaIP = 1e-05,
  Bootstrapping = TRUE,
  M = 100,
  seed = 777,
  Plot = TRUE,
  show_adjustments = FALSE
)
}
\arguments{
\item{dat}{Data frame. Must have header with at least the \code{xbeta}, \code{xse}, \code{ybeta} and \code{yse} columns present.}

\item{snp_col}{Name of column with SNP IDs.}

\item{xbeta_col}{Required name of column with effects on the incidence trait.}

\item{xse_col}{Required name of column with standard errors of \code{xbeta}.}

\item{xp_col}{Name of column with p-value for \code{xbeta} (optional). If not given, It will be inferred from \code{xbeta} and \code{xse}.}

\item{ybeta_col}{Required name of column with unadjusted effects on the prognosis trait.}

\item{yse_col}{Required name of column with standard errors of \code{ybeta}.}

\item{yp_col}{Name of column with p-value for \code{ybeta} (optional). If not given, It will be inferred from \code{ybeta} and \code{yse}.}

\item{xp_thresh}{p-value threshold for SNP-incidence associations. Effects with p-values larger than \code{xp.thresh} will be excluded prior to fitting the main model-based clustering.}

\item{init_pi}{initial value for the weight of the mixture component that represents the cluster of SNPs affecting x only.}

\item{init_sigmaIP}{initial value for the covariance between x and y.}

\item{Bootstrapping}{Logical, if TRUE estimate the standard error of the adjustment factor using the Bootstrap method.}

\item{M}{Number of bootstrap samples drawn to estimate the standard error of the adjustment factor.}

\item{seed}{Random number seed used for drawing the bootstrap samples.}

\item{Plot}{Logical, if TRUE (the default), calling the function should plot the final clusters.}

\item{show_adjustments}{Logical indicating whether to show adjusted effects of the given SNPs in the outputs.}
}
\value{
List of the following:
\itemize{
\item est: estimated adjusted associations, their standard errors and p-values (only if \code{show_adjustments} is TRUE).
\item b: The estimated slope (adjustment factor).
\item bse: Standard error of the estimated slope.
\item b_CI: 95\\% confidence interval of the adjustment factor.
\item pi: Estimated probability of the mixture component of SNPs affecting only incidence.
\item entropy: The entropy of the estimated clusters.
\item plot: Generated plot of the SlopeHunter fitted model.
\item Fit: a Data frame  summarising the fitted model-based clustering with the following columns:
\itemize{
\item cluster: cluster of the variants defined as follows:
\itemize{
\item \code{Hunted} = assigned to the cluster of SNPs affecting only incidence.
\item \code{Pleiotropic} = assigned to the cluster affecting both incidence and prognosis - i.e. variants that affect incidence and have direct effect on prognosis.
}
\item pt and p0: membership probabilities of the variants for the hunted and pleiotropic clusters respectively.
\item associations of variants with x and y, their standard errors and p-values.
}
\item iter: Number of the EM algorithm's iterations.
\item Bts.est: Details on the bootstrap estimate of the standard error of the adjustment factor, if \code{Bootstrapping} is TRUE.
}
}
\description{
Estimate collider bias
}
