% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD_prune.R
\name{LD_prune}
\alias{LD_prune}
\title{Perform LD pruning on SNP data}
\usage{
LD_prune(
  dat,
  clump_kb = 250,
  clump_r2 = 0.1,
  Random = TRUE,
  clump_p1 = 1,
  local = FALSE,
  ref_pop = "EUR",
  ref_bfile,
  seed = 77777
)
}
\arguments{
\item{dat}{Output from \code{harmonise_effects}. Must have a SNP name column (SNP).}

\item{clump_kb}{Clumping window, default is \code{250}.}

\item{clump_r2}{Clumping r-squared threshold, default is \code{0.1}.}

\item{Random}{Logical, if \code{TRUE} (the default), SNPs will be randomly pruned. Otherwise, based on p-values.}

\item{clump_p1}{Clumping sig level for index SNPs, default is \code{1}.}

\item{local}{Logical, if \code{FALSE} (the default), the MRC-IEU API 'http://gwas-api.mrcieu.ac.uk/'
will be used for clumping. Otherwise, your local machine will be used for clumping given that you
provide a bed/bim/fam LD reference dataset.}

\item{ref_pop}{Super-population to use as reference panel at the API (when \code{local} is FALSE). Default = \code{"EUR"}.}

\item{ref_bfile}{Path to the bed/bim/fam LD reference (e.g. "1kg.v3/EUR" for local 1000 EUR ref. population file). If \code{local}=TRUE, then this should be provided.}

\item{seed}{Random number seed for random pruning}
}
\value{
Data frame
}
\description{
Uses PLINK clumping method ('--clump' command), where a greedy search algorithm is implemented to randomly select a variant
(or the variant with the lowest p-value, if a user wish to), referred to as the index SNP, and remove all variants within a
certain kb distance in linkage disequilibrium with the index SNP, based on an r-squared threshold from the 1000 Genomes
reference panel phase 3 data. Then repeats until no variants are left.
}
