% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SH-utils.R
\name{plot.SH}
\alias{plot.SH}
\title{Plotting model for Slope-Hunter clustering}
\usage{
\method{plot}{SH}(
  x,
  what = c("clusters", "classification", "uncertainty", "density"),
  xlab = NULL,
  ylab = NULL,
  addEllipses = TRUE,
  main = FALSE,
  ...
)
}
\arguments{
\item{x}{Output from \code{slopehunter}.}

\item{what}{A string specifying the type of graph requested. Available choices are:
"clusters": showing clusters. The plot can display membership probabilities of each variable (e.g. SNP) to the target cluster (G1) by hovering over the points.
"classification": A plot showing point assigned to each cluster (class).
"uncertainty": A plot of classification uncertainty.
"density": A plot of estimated density.}

\item{xlab}{Optional label for the x-axis in case of "classification", "uncertainty", or "density" plots.}

\item{ylab}{Optional label for the y-axis in case of "classification", "uncertainty", or "density" plots.}

\item{addEllipses}{A logical indicating whether or not to add ellipses with axes corresponding to the within-cluster covariances in case of "classification" or "uncertainty" plots.}

\item{main}{A logical or NULL indicating whether or not to add a title to the plot identifying the type of plot drawn in case of "classification", "uncertainty", or "density" plots.}

\item{...}{Other graphics parameters.}
}
\description{
Plotting model for Slope-Hunter clustering
}
