% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_incidence.R
\name{read_incidence}
\alias{read_incidence}
\title{Read incidence data}
\usage{
read_incidence(
  filename,
  gz = TRUE,
  sep = " ",
  snp_col = "SNP",
  beta_col = "BETA",
  se_col = "SE",
  pval_col = "PVAL",
  eaf_col = "EAF",
  effect_allele_col = "EA",
  other_allele_col = "OA",
  gene_col = "GENE",
  chr_col = "CHR",
  pos_col = "POS",
  min_pval = 1e-200,
  log_pval = FALSE
)
}
\arguments{
\item{filename}{Filename. Must have header with at least the \code{SNP}, \code{beta}, \code{se} and \code{EA}columns present.}

\item{gz}{whether the given data file is compressed with the gzip (\code{.gz} file). The default is \code{TRUE}.}

\item{sep}{Specify delimeter in file if \code{gz} is FALSE. The default is a space, i.e. \code{" "}.}

\item{snp_col}{Required name of column with SNP rs IDs. The default is \code{"SNP"}.}

\item{beta_col}{Required name of column with effect sizes. The default is \code{"BETA"}.}

\item{se_col}{Required name of column with standard errors. The default is \code{"SE"}.}

\item{pval_col}{Name of column with p-value (optional). The default is \code{"PVAL"}. It will be Inferred when possible from beta and se.}

\item{eaf_col}{Name of column with effect allele frequency (optional). The default is \code{"EAF"}.}

\item{effect_allele_col}{Required for harmonisation. Name of column with effect allele. Must be "A", "C", "T" or "G". The default is \code{"EA"}.}

\item{other_allele_col}{Required for harmonisation. Name of column with non-effect allele. Must be "A", "C", "T" or "G". The default is \code{"OA"}.}

\item{gene_col}{Optional column for gene name. The default is \code{"GENE"}.}

\item{chr_col}{Optional column for chromosome number. The default is \code{"CHR"}.}

\item{pos_col}{Optional column for SNP position. The default is \code{"POS"}.}

\item{min_pval}{Minimum allowed p-value. The default is \code{1e-200}.}

\item{log_pval}{The p-value is -log10(P). The default is \code{FALSE}.}
}
\value{
data frame
}
\description{
Reads in incidence data. Checks and organises columns for use with the Slope-Hunter analyses. Infers p-values when possible from beta and se.
}
