% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHadj.R
\name{SHadj}
\alias{SHadj}
\title{Correct index event bias for new data}
\usage{
SHadj(
  x,
  dat,
  snp_col = "SNP",
  xbeta_col = "BETA.incidence",
  xse_col = "SE.incidence",
  ybeta_col = "BETA.prognosis",
  yse_col = "SE.prognosis"
)
}
\arguments{
\item{x}{an pbject of the class SH obtained from the \code{slopehunter} function.}

\item{dat}{A data.frame with harmonised effects and alleles, formatted using the \code{harmonise_effects} function.}

\item{snp_col}{Name of column with SNP IDs.}

\item{xbeta_col}{Required name of column with effects on the incidence trait.}

\item{xse_col}{Required name of column with standard errors of \code{xbeta}.}

\item{ybeta_col}{Required name of column with unadjusted effects on the prognosis trait.}

\item{yse_col}{Required name of column with standard errors of \code{ybeta}.}
}
\value{
data.frame with adjusted estimates
}
\description{
Correct index event bias for new data
}
