% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SH-utils.R
\name{shclust}
\alias{shclust}
\title{Implement the EM algorithm for the SlopeHunter model-based clustering}
\usage{
shclust(gwas, pi0, sxy1)
}
\arguments{
\item{gwas}{a data frame with columns: xbeta; xse; ybeta; yse.}

\item{pi0}{initial value for the weight of the mixture component that represents the cluster of SNPs affecting x only.}

\item{sxy1}{initial value for the covariance between x and y.}
}
\value{
EM fit for SlopeHunter estimator
}
\description{
Implement the EM algorithm for the SlopeHunter model-based clustering
}
