% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopehunter.R
\name{slopehunter}
\alias{slopehunter}
\title{Estimate index even bias}
\usage{
slopehunter(
  dat,
  snp_col = "SNP",
  xbeta_col = "BETA.incidence",
  xse_col = "SE.incidence",
  xp_col = "Pval.incidence",
  ybeta_col = "BETA.prognosis",
  yse_col = "SE.prognosis",
  yp_col = "Pval.prognosis",
  comp.size = seq(0.03, 0.1, 0.01),
  xp.thresh = 0.1,
  coef.diff = 1,
  seed = 2019,
  correct.reg.dill = TRUE,
  show_adjustments = FALSE,
  Plot = TRUE
)
}
\arguments{
\item{dat}{Data frame. Must have header with at least the `xbeta`, `xse`, `ybeta` and `yse` columns present.}

\item{snp_col}{Name of column with SNP IDs.}

\item{xbeta_col}{Required name of column with effects on the incidence trait.}

\item{xse_col}{Required name of column with standard errors of \code{xbeta}.}

\item{xp_col}{Name of column with p-value for \code{xbeta} (optional). If not given, It will be inferred from \code{xbeta} and \code{xse}.}

\item{ybeta_col}{Required name of column with unadjusted effects on the prognosis trait.}

\item{yse_col}{Required name of column with standard errors of \code{ybeta}.}

\item{yp_col}{Name of column with p-value for \code{ybeta} (optional). If not given, It will be inferred from \code{ybeta} and \code{yse}.}

\item{comp.size}{Vector of complementary set sizes as a proportion of the noisy SNPs (neither affecting incidence nor subsequent traits).
The function will iteratively fit the `slopehunter` model using each of these sizes.}

\item{xp.thresh}{p-value threshold for SNP-incidence associations. Effects with p-values higher than th\code{xp.thresh} will be dropped from the analysis.}

\item{coef.diff}{A multiplier (coefficient) of the minimum standard deviation of the fitted clusters (say 'sd.min'). The product of
\code{coef.diff} with 'sd.min' giving a ceiling for the Euclidean distance between the two cluster means of an accepted model fit.
It is used for trimming the set of candidate models corresponding to each size in \code{comp.size}. The default value is 1. Larger values result in
larger set of possible model solutions.}

\item{seed}{Random number seed for the cluster-based models and weighted selections from the noisy SNPs}

\item{correct.reg.dill}{Logical indicating whether to correct for regression dillution using the Hedges Olkin method (the default is TRUE).}

\item{show_adjustments}{Logical indicating whether to show adjusted effects of the given SNPs in the outputs.}

\item{Plot}{Logical, if TRUE (the default), calling the function should plot the final clusters}
}
\value{
List of the following:
\itemize{
\item Estimates: Data frame with:
\itemize{
\item status of the variants defined as follows:
\itemize{
\item `G4-Drop`: assigned to the noisy cluster (excluded from the analysis).
\item `G3-Drop` = assigned to the cluster affecting prognosis with no effect on incidence (excluded from the analysis).
\item `G1` = the target cluster.
\item `G2` = assigned to the cluster affecting both incidence and prognosis - i.e. variants with pleiotropic effect.
\item `G1-Retained` = Retained from the excluded G4 cluster and assigned to the target cluster.
\item `G2-Retained` = Retained from the excluded G4 cluster and assigned to the cluster affecting both incidence and prognosis traits).
}
\item membership probabilities of the vaiants based on which they are assigned to the target cluster (obviously these are calculated for only the variants included in the analysis)
\item adjusted effect sizes, their standard errors and p-values (only if \code{show_adjustments} is TRUE).
}
\item Iterations: Details of the iterations.
\Item Model: the fitted model.
\item b.raw: The estimated slope (adjustment factor).
\item Sh.b: The estimated slope adjusted for regression dilution, if `correct.reg.dill` is TRUE.
\item bse: Standard error of the estimated slope.
}
}
\description{
Estimate index even bias
}
