% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SH-utils.R
\name{plot.SH}
\alias{plot.SH}
\title{Plotting model for Slope-Hunter clustering}
\usage{
\method{plot}{SH}(
  x,
  what = c("clusters", "classification", "uncertainty", "density"),
  dimens = NULL,
  xlab = NULL,
  ylab = NULL,
  addEllipses = TRUE,
  main = FALSE,
  ...
)
}
\arguments{
\item{x}{Output from \code{slopehunter}.}

\item{what}{A string specifying the type of graph requested. Available choices are:
"clusters": showing clusters. The plot can display membership probabilities of each variable (e.g. SNP) to the target cluster (G1) by hovering over the points.
"classification": A plot showing point assigned to each cluster (class).
"uncertainty": A plot of classification uncertainty.
"density": A plot of estimated density.}
}
\description{
Plotting model for Slope-Hunter clustering
}
